+++ pkg/reconciler/githubsource/doc.go
@@ -15,8 +15,4 @@ limitations under the License.
 */
 
 // Package githubsource implements a GitHubSource controller.
-// +kubebuilder:rbac:groups=serving.knative.dev,resources=services,verbs=get;list;watch;create;update;patch;delete
-// +kubebuilder:rbac:groups=sources.eventing.knative.dev,resources=githubsources,verbs=get;list;watch;create;update;patch;delete
-// +kubebuilder:rbac:groups=sources.eventing.knative.dev,resources=githubsources/status,verbs=get;update;patch
-// +kubebuilder:rbac:groups=core,resources=secrets,verbs=get;list;watch
 package githubsource
